maybe fix test not being able to compile under windows
